refactor: Split regen ledger into multiple go modules (#322)
* set up app go module

* add standalone data module

* add ecocredit go module

* WIP:add types,testutil,orm,group modules

* WIP remove dep on app in types/testutil

* wip no dep cycles detected

* apply reviewer comments

* remove comments

* move math to types/math

* types/testutil-orm cycle dep detected

* move codec.proto, codec.pb.go, model.go from testdata to orm

* move codec.proto codec.pb.go model.go to orm/testdata

* protocgen error

* trying resolve error

* fix typos

* update makefile and generate codec.pb.go in orm/testdata

* move tx.* files to x/group/testdata

* gen tx.pb.go in x/group/testdata

* run go mod tidy in every module, found some issues

* move client/msg_service.go to types/module/msg_service.go and move util/utils.go to x/ecocredit/util/utils.go

* WIP fix tests

* trying to fix tests

* add build experimental in modules_test

* fix build experimental in modules_test.go

* cleanup

* apply reviewers comments

* Update x/group/client/testsuite/tx.go


* Update x/data/client/testsuite/suite.go


* Update x/group/client/testsuite/tx.go


* Update types/testutil/network/suite.go
