+++ orm/testdata/codec.proto
@@ -4,7 +4,7 @@ package testdata;
 
 import "gogoproto/gogo.proto";
 
-option go_package = "github.com/regen-network/regen-ledger/testutil/testdata";
+option go_package = "github.com/regen-network/regen-ledger/orm/testdata";
 
 message GroupInfo {
   uint64 group_id = 1;
