+++ scripts/protocgen.sh
@@ -73,8 +73,8 @@ google/protobuf/any.proto=github.com/cosmos/cosmos-sdk/codec/types:. \
 done
 
 # generate codec/testdata proto code
-buf protoc -I "proto" -I "third_party/proto" -I "testutil/testdata" --gocosmos_out=plugins=interfacetype+grpc,\
-Mgoogle/protobuf/any.proto=github.com/cosmos/cosmos-sdk/codec/types:. ./testutil/testdata/*.proto
+buf protoc -I "proto" -I "third_party/proto" -I "x/group/testdata" --gocosmos_out=plugins=interfacetype+grpc,\
+Mgoogle/protobuf/any.proto=github.com/cosmos/cosmos-sdk/codec/types:. ./x/group/testdata/*.proto
 
 # move proto files to the right places
 cp -r github.com/regen-network/regen-ledger/* ./
