+++ x/ecocredit/server/helpers.go
@@ -10,7 +10,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
-	"github.com/regen-network/regen-ledger/math"
+	"github.com/regen-network/regen-ledger/types/math"
 )
 
 func getDecimal(store sdk.KVStore, key []byte) (*apd.Decimal, error) {
