+++ x/ecocredit/server/testsuite/suite.go
@@ -3,7 +3,7 @@ package testsuite
 import (
 	"context"
 
-	"github.com/regen-network/regen-ledger/testutil"
+	"github.com/regen-network/regen-ledger/types/testutil"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/suite"
