+++ x/group/client/testsuite/query.go
@@ -1,4 +1,3 @@
-// +build experimental
 
 package testsuite
 
@@ -6,9 +5,8 @@ import (
 	"fmt"
 	"strconv"
 
+	"github.com/regen-network/regen-ledger/types/testutil/cli"
 	tmcli "github.com/tendermint/tendermint/libs/cli"
-
-	"github.com/regen-network/regen-ledger/testutil/cli"
 	"github.com/regen-network/regen-ledger/x/group"
 	"github.com/regen-network/regen-ledger/x/group/client"
 )
